주문서 생성 후->요청상품추가(가격흥정중) 또는 바로 결제(결제완료) 선택 (#17)

+++ app/src/main/java/com/example/graduationproject/Api/Response/MartInfoResponse.kt
@@ -10,7 +10,7 @@ data class MartInfoResponse(
     @SerializedName("martNumber")
     val martNumber:String,
     @SerializedName("requestYn")
-    val requestYn:Long,
+    val requestYn:Int,
     @SerializedName("martAddress")
     val martAddress: String
 )
